requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asarPy==1.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
gevent==25.9.1
greenlet==3.2.4
idna==3.10
magic-filter==1.0.12
python-telegram-bot==21.6
multidict==6.6.4
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dotenv==1.1.1
reportlab==4.4.4
requests==2.32.5
setuptools==80.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
zope.event==6.0
zope.interface==8.0.1
